# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ItehProject API documentation",
        "version": "0.0.1",
        "title": "ItehProject API",
        "contact": {},
        "license": {}
    },
    "host": "localhost:9060",
    "basePath": "/",
...
    "paths": {
...
        "/api/_search/exams": {
            "get": {
                "tags": [
                    "exam-resource"
                ],
                "summary": "searchExams",
                "operationId": "searchExamsUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number of the requested page",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Size of a page",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ExamDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/_search/student-exams": {
            "get": {
                "tags": [
                    "student-exam-resource"
                ],
                "summary": "searchStudentExams",
                "operationId": "searchStudentExamsUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number of the requested page",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Size of a page",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/StudentExamDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/apply-for-exam": {
            "put": {
                "tags": [
                    "student-exam-resource"
                ],
                "summary": "Apply for exam",
                "description": "Creates an exam application for the next period identified by examId. Returns the page data again",
                "operationId": "applyForExamUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "exam",
                        "in": "query",
                        "description": "exam",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ExamApplicationPageData"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/cancel-exam-application": {
            "put": {
                "tags": [
                    "student-exam-resource"
                ],
                "summary": "Cancel exam application",
                "description": "Cancels the exam application identified by examId. Returns the page data again.",
                "operationId": "cancelExamApplicationUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "exam",
                        "in": "query",
                        "description": "exam",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ExamApplicationPageData"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/exam-application-data": {
            "get": {
                "tags": [
                    "student-exam-resource"
                ],
                "summary": "Return data for exam applications page",
                "description": "Returns all data for the exam application page for the logged in student, which is all exams enrolled, not passed, not applied for and exams applied for the next period.",
                "operationId": "examApplicationPageDataUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ExamApplicationPageData"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/exams": {
            "get": {
                "tags": [
                    "exam-resource"
                ],
                "summary": "getAllExams",
                "operationId": "getAllExamsUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number of the requested page",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Size of a page",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ExamDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "exam-resource"
                ],
                "summary": "createExam",
                "operationId": "createExamUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "examDTO",
                        "description": "examDTO",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ExamDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ExamDTO"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "exam-resource"
                ],
                "summary": "updateExam",
                "operationId": "updateExamUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "examDTO",
                        "description": "examDTO",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ExamDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ExamDTO"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/exams/{id}": {
            "get": {
                "tags": [
                    "exam-resource"
                ],
                "summary": "getExam",
                "operationId": "getExamUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ExamDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "exam-resource"
                ],
                "summary": "deleteExam",
                "operationId": "deleteExamUsingDELETE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/student-exams/{id}": {
            "get": {
                "tags": [
                    "student-exam-resource"
                ],
                "summary": "getStudentExam",
                "operationId": "getStudentExamUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/StudentExamDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "student-exam-resource"
                ],
                "summary": "deleteStudentExam",
                "operationId": "deleteStudentExamUsingDELETE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ExamApplicationPageData": {
            "type": "object",
            "properties": {
                "appliedExams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ExamDTO"
                    }
                },
                "exams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ExamDTO"
                    }
                },
                "nextPeriod": {
                    "$ref": "#/definitions/ExamPeriodDTO"
                }
            }
        },
        "ExamDTO": {
            "type": "object",
            "required": [
                "time"
            ],
            "properties": {
                "courseId": {
                    "type": "integer",
                    "format": "int64"
                },
                "courseName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "periodId": {
                    "type": "integer",
                    "format": "int64"
                },
                "periodName": {
                    "type": "string"
                },
                "time": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "ExamPeriodDTO": {
            "type": "object",
            "required": [
                "endDate",
                "name",
                "startDate"
            ],
            "properties": {
                "endDate": {
                    "type": "string",
                    "format": "date"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 100
                },
                "startDate": {
                    "type": "string",
                    "format": "date"
                },
                "yearId": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
...
        "StudentExamDTO": {
            "type": "object",
            "properties": {
                "attended": {
                    "type": "boolean"
                },
                "evaluatedByFullName": {
                    "type": "string"
                },
                "evaluatedById": {
                    "type": "integer",
                    "format": "int64"
                },
                "examId": {
                    "type": "integer",
                    "format": "int64"
                },
                "examName": {
                    "type": "string"
                },
                "grade": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 6.0,
                    "maximum": 10.0,
                    "exclusiveMinimum": false,
                    "exclusiveMaximum": false
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "studentFullName": {
                    "type": "string"
                },
                "studentId": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
...
    }
}